requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
pypdf==4.3.1
google-genai==1.66.0
python-dotenv==1.0.1
snowballstemmer==2.2.0
scikit-learn==1.5.2
pymupdf==1.24.10
pytesseract==0.3.10
pillow==10.4.0
